move a few things
