+++ src/hypofuzz/utils.py
@@ -1,9 +1,12 @@
 import math
-from typing import Any
+import threading
+from typing import Any, Generic, TypeVar
 
 FUZZJSON_INF = "hypofuzz-inf-a928fa52b3ea4a9a9af36ccef7c6cf93"
 FUZZJSON_NINF = "hypofuzz-ninf-a928fa52b3ea4a9a9af36ccef7c6cf93"
 FUZZJSON_NAN = "hypofuzz-nan-a928fa52b3ea4a9a9af36ccef7c6cf93"
+
+T = TypeVar("T")
 
 
 def convert_to_fuzzjson(value: Any) -> Any:
@@ -28,3 +31,23 @@ def convert_to_fuzzjson(value: Any) -> Any:
         return value
     else:
         raise ValueError(f"unknown type {type(value)} ({value!r})")
+
+
+def lerp(a: float, b: float, t: float) -> float:
+    return (1 - t) * a + t * b
+
+
+# hypothesis.utils.dynamicvaraible, but without the with_value context manager.
+# Essentially just a reference to a value.
+class Value(Generic[T]):
+    def __init__(self, default: T) -> None:
+        self.default = default
+        self.data = threading.local()
+
+    @property
+    def value(self) -> T:
+        return getattr(self.data, "value", self.default)
+
+    @value.setter
+    def value(self, value: T) -> None:
+        self.data.value = value
